feat: add support for Jotai V2 API
