+++ package.json
@@ -8,16 +8,14 @@
   },
   "sideEffects": false,
   "source": "./src/index.ts",
+  "types": "./dist/index.d.ts",
+  "main": "./dist/index.cjs.js",
   "exports": {
     "./package.json": "./package.json",
     ".": {
       "import": {
         "types": "./dist/index.d.mts",
         "default": "./dist/index.esm.mjs"
-      },
-      "module": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.esm.js"
       },
       "default": {
         "types": "./dist/index.d.ts",
@@ -29,19 +27,12 @@
         "types": "./dist/*.d.mts",
         "default": "./dist/*.esm.mjs"
       },
-      "module": {
-        "types": "./dist/*.d.ts",
-        "default": "./dist/*.esm.js"
-      },
       "default": {
         "types": "./dist/*.d.ts",
         "default": "./dist/*.cjs.js"
       }
     }
   },
-  "types": "./dist/index.d.ts",
-  "main": "./dist/index.cjs.js",
-  "module": "./dist/index.esm.js",
   "files": [
     "dist",
     "package.json"
@@ -105,7 +96,7 @@
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.3.1",
     "jest-watch-typeahead": "^2.2.1",
-    "jotai": "^1.12.1",
+    "jotai": "^1.13.0",
     "lint-staged": "^13.1.0",
     "prettier": "^2.8.1",
     "react": "^18.2.0",
@@ -117,8 +108,8 @@
     "typescript": "^4.9.4"
   },
   "peerDependencies": {
-    "jotai": ">=1.8.6 <2.x",
-    "react": ">=16.8"
+    "jotai": ">=1.11.0",
+    "react": ">=17.0.0"
   },
   "dependencies": {
     "@redux-devtools/extension": "^3.2.3"
