--- src/constants.ts
@@ -1,7 +1,0 @@
-// These are internal values exported from `jotai/core/store` but are not exposed via the public API for us to use
-// See https://github.com/pmndrs/jotai/blob/f0048c15b6ba4271bf26e18b0e5bb59b83db33d7/src/core/store.ts#L138-L144
-export const DEV_SUBSCRIBE_STATE = 'n';
-export const DEV_GET_MOUNTED_ATOMS = 'l';
-export const DEV_GET_ATOM_STATE = 'a';
-export const DEV_GET_MOUNTED = 'm';
-export const RESTORE_ATOMS = 'h';
