+++ src/index.ts
@@ -1,7 +1,1 @@
-export {
-  useAtomsSnapshot,
-  useGotoAtomsSnapshot,
-  useAtomsDebugValue,
-  useAtomDevtools,
-  useAtomsDevtools,
-} from './utils/';
+export * from './utils';
