+++ src/utils/useGotoAtomsSnapshot.ts
@@ -1,57 +1,15 @@
-import { useCallback, useContext } from 'react';
-import { SECRET_INTERNAL_getScopeContext as getScopeContext } from 'jotai';
-import type { Atom } from 'jotai';
-import { DEV_SUBSCRIBE_STATE, RESTORE_ATOMS } from '../constants';
+import { useCallback } from 'react';
+import { useStore } from 'jotai/react';
+import type { AtomsSnapshot, Options } from './types';
 
-type Scope = NonNullable<Parameters<typeof getScopeContext>[0]>;
-type AnyAtomValue = unknown;
-type AnyAtom = Atom<AnyAtomValue>;
-type AtomsValues = Map<AnyAtom, AnyAtomValue>; // immutable
-type AtomsDependents = Map<AnyAtom, Set<AnyAtom>>; // immutable
-type AtomsSnapshot = Readonly<{
-  values: AtomsValues;
-  dependents: AtomsDependents;
-}>;
-
-/**
- * @deprecated use object snapshot instead
- */
-type DeprecatedIterableSnapshot = Iterable<readonly [AnyAtom, AnyAtomValue]>;
-
-export function useGotoAtomsSnapshot(scope?: Scope) {
-  const ScopeContext = getScopeContext(scope);
-  const { s: store, w: versionedWrite } = useContext(ScopeContext);
-
+export function useGotoAtomsSnapshot(options?: Options) {
+  const store = useStore(options);
   return useCallback(
-    (snapshot: AtomsSnapshot | DeprecatedIterableSnapshot) => {
-      if (!store[DEV_SUBSCRIBE_STATE]) return;
-
-      const restoreAtoms = (
-        values: Iterable<readonly [AnyAtom, AnyAtomValue]>,
-      ) => {
-        if (versionedWrite) {
-          versionedWrite((version) => {
-            store[RESTORE_ATOMS](values, version);
-          });
-        } else {
-          store[RESTORE_ATOMS](values);
-        }
-      };
-      if (isIterable(snapshot)) {
-        if (__DEV__) {
-          console.warn(
-            'snapshot as iterable is deprecated. use an object instead.',
-          );
-        }
-        restoreAtoms(snapshot);
-        return;
+    (snapshot: AtomsSnapshot) => {
+      if (store.dev_restore_atoms) {
+        store.dev_restore_atoms(snapshot.values);
       }
-      restoreAtoms(snapshot.values);
     },
-    [store, versionedWrite],
+    [store],
   );
 }
-
-const isIterable = (item: any): item is Iterable<any> => {
-  return typeof item[Symbol.iterator] === 'function';
-};
